### 2020-10-29 #2796 @whedon
[PRE REVIEW]: SHED: Streaming Heterogeneous Event Data Tracking with Provenance
**Submitting author:** @CJ-Wright (<a href="http://orcid.org/0000-0003-2522-7028">Christopher Wright</a>)
**Repository:** <a href="https://github.com/xpdAcq/SHED" target ="_blank">https://github.com/xpdAcq/SHED</a>
**Version:** 0.7.5
