### 2020-11-03 #2796 @katyhuff
* @remram44 : Remi Rampin, are you able to and interested in taking on this review for JOSS?
* This submission is in need of expertise in reproducibility tooling as well as familiarity with the systems and languages involved (primarily python).
* Title: SHED: Streaming Heterogeneous Event Data Tracking with Provenance
* Summary: The SHED package provides a python based software framework for real time analysis of streaming data with provenance. It is highly flexible, allowing analysis pipelines to be built for any variety of experimental data that is arriving in a time series. It allows straightforward serialization of the pipeline, and the analysis results, for storage in databases with provenance information that allows analyses later to be pulled from the databases, adapted if desired, and rerun.
* Article Proof: https://raw.githubusercontent.com/openjournals/joss-papers/joss.02796/joss.02796/10.21105.joss.02796.pdf
* Submitting author: @CJ-Wright (Christopher Wright)
* Repository: https://github.com/xpdAcq/SHED
* Version: 0.7.5
