### 2021-03-14 #2796 @arfon
:wave: @julianpistorius @mbdemoraes – would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under review is [_SHED: Streaming Heterogeneous Event Data Tracking with Provenance_](https://github.com/xpdAcq/SHED).
