### 2021-03-17 #3119 @whedon
[REVIEW]: SHED: Streaming Heterogeneous Event Data Tracking with Provenance
**Submitting author:** @CJ-Wright (<a href="http://orcid.org/0000-0003-2522-7028">Christopher Wright</a>)
**Repository:** https://github.com/xpdAcq/SHED
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.7.5
