### 2021-03-29 #3119 @remram44
I have taken a first look at the repositories. Some notes:

* **General**: `shed` is taken on PyPI, by https://github.com/Zac-HD/shed. It might be beneficial to use a unique name, such as `shed-data`, `shed-streaming`, etc to avoid conflicting in people's environments. Should the other `shed` decide to publish on conda-forge, things will get weird (I was surprised this was not discussed there https://github.com/conda-forge/staged-recipes/pull/4804).
* **Installation instructions**: Installing from the repo is not documented, only using the package from conda-forge. I think this should be added if people are to contribute or fork the project.
  * I was able to install from the `requirements/` files, but noted that `pip.txt` is empty, which is probably not intended?
* **Automated tests**: the tests don't run for me, possibly because `pytest-tornado` is the wrong version. I note that you use an additional Conda channel on Travis, [`nsls2forge`](https://anaconda.org/nsls2forge), which seems to be your own.
* **Automated tests**: you do use Travis for CI, you might want to include a link to the Travis build page (for example their badge) https://travis-ci.org/github/xpdAcq/SHED
* **Functionality documentation**: The documentation page https://xpdacq.github.io/SHED/shed.html shows up as empty except for headers. It looks like documentation exists and there is just something wrong with autodoc.
* **Documentation**: The rest of the documentation is pretty sparse. There is little explanation of the purpose of the software; maybe some text can be copied from the paper.
* **Tutorials**: They build on NSLS-II data standards and the `bluesky` framework, which might not be widely known/used outside of the Brookhaven National Laboratory. It also uses `rapidz`, which is a fork of `streamz` from the same group. The paper mentions `streamz` but `rapidz` is actually what's used throughout the package, except for one example file (`average.py`).
* **Paper**: The paper looks good, I sent some nit-picks at https://github.com/xpdAcq/SHED/pull/202
* **References** / **State of the field**: Maybe more references could be added to modern streaming dataflow system/models such as Timely dataflows which are popular recently? In addition, I think more details on what SHED adds compared to `dask` and `streamz` (that it builds upon) or [`ray`](https://rise.cs.berkeley.edu/projects/ray/) might be interesting?

I will look into the code next.
