### 2021-04-22 #3119 @sbillinge
Thanks for your hard work on this @remram44 .  You are right, that somehow I wasn't watching the repo and I didn't get a notification so I missed all your hard work.  I am now watching.  Sorry, this is the first time I did this process at JOSS.   We will work on the suggestions. @CJ-Wright Looping you in.  I will do what I can.  There are some things we may want to chat about.
