### 2021-06-27 #3119 @sbillinge
Dear Arfon,

I would be willing to help with reviewing, but this is my paper, so it is
not so appropriate for me to review it!

S

On Sun, Jun 27, 2021 at 4:59 AM Arfon Smith ***@***.***>
wrote:

> 👋 @alesarrett <https://github.com/alesarrett> – you be willing to review
> this submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> We already have one reviewer ***@***.*** <https://github.com/remram44>)
> and are looking to find a second.
>
> The package under review is *SHED: Streaming Heterogeneous Event Data
> Tracking with Provenance* <https://github.com/xpdAcq/SHED>, a Python tool
> for streaming data from event models, while tracking data provenance and
> workflow execution.
>
