### 2021-08-07 #3119 @remram44
Thanks @arfon, I must say I haven't been prioritizing this since we are missing the second reviewer. I was also hoping some of my points in https://github.com/openjournals/joss-reviews/issues/3119#issuecomment-809781527 would be addressed, but that doesn't block me from running the software so I will do that ASAP.
