### 2021-10-05 #3119 @alesarrett
> wave @alesarrett – you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> We already have one reviewer (@remram44) and are looking to find a second.
>
> The package under review is [_SHED: Streaming Heterogeneous Event Data Tracking with Provenance_](https://github.com/xpdAcq/SHED), a Python tool for streaming data from event models, while tracking data provenance and workflow execution.

Sorry to be so late in replying.
Anyway, I'm not able to hep with this review, sorry.
