### 2021-11-11 #3119 @arfon
Documenting some additional activity on my part here: I've just emailed @CJ-Wright & @sbillinge to ask them:

1. To respond to @remram44's [initial review feedback above](https://github.com/openjournals/joss-reviews/issues/3119#issuecomment-8097815271)
2. To assist with finding a second reviewer.
