### 2022-02-18 #3119 @arfon
@CJ-Wright & @sbillinge – it's close to 8 months since we've heard back from you about this submission. At this point I'm assuming you're no longer interested in pursuing this submission in JOSS.

If we don't hear back from you in the next week, I'll reject this submission due to lack of activity.
