### 2022-02-22 #3119 @sbillinge
Sorry about this Arfon.  My student CJ left the group with things in this
state.  I will check with him how we wants to proceed but he will have to
commit to spending some time on it for it to work out.

S

On Fri, Feb 18, 2022 at 3:34 PM Arfon Smith ***@***.***>
wrote:

> @CJ-Wright <https://github.com/CJ-Wright> & @sbillinge
> <https://github.com/sbillinge> – it's close to 8 months since we've heard
> back from you about this submission. At this point I'm assuming you're no
> longer interested in pursuing this submission in JOSS.
>
> If we don't hear back from you in the next week, I'll reject this
> submission due to lack of activity.
>
