+++ requirements/doc.txt
@@ -1,4 +1,3 @@
 sphinx
 sphinx_rtd_theme
 doctr
-rever
