+++ requirements/build.txt
@@ -1,2 +1,3 @@
 python
 setuptools
+pip
