### 2022-02-24 d8b6977
BLD: add install.sh
