### 2022-02-24 9407bab
DOC: add installation from source
