### 2022-02-24 ecc3722
BLD: add pip install
