### 2022-02-24 b85dc87
DOC: delete tests package and module contents
