+++ requirements/run.txt
@@ -1,3 +1,4 @@
+python
 streamz_ext
 numpy
 scipy
@@ -8,4 +9,5 @@ event-model
 networkx
 rapidz
 xonsh
+dask==2.20.0
 dask-jobqueue
