+++ examples/average.py
@@ -2,7 +2,7 @@ from bluesky.run_engine import RunEngine
 from ophyd.sim import hw
 from bluesky.plans import count
 from shed.translation import FromEventStream, ToEventStream
-from streamz import Stream
+from rapidz import Stream
 from bluesky.callbacks import LivePlot
 import matplotlib.pyplot as plt
 
