### 2022-02-28 5922203
MNT: rename the project and package to shed-streaming.
