+++ setup.py
@@ -1,10 +1,10 @@
 from setuptools import setup, find_packages
 
 setup(
-    name='shed',
+    name='shed_streaming',
     version='0.7.5',
     packages=find_packages(),
     # description='data processing module',
     zip_safe=False,
-    url='http:/github.com/xpdAcq/shed',
+    url='http:/github.com/xpdAcq/shed-streaming',
 )
