### 2022-02-28 0fb82b9
MNT: correct the wrong names in the docs
