+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, find_packages
 
 setup(
     name='shed_streaming',
-    version='0.7.5',
+    version='0.8.0',
     packages=find_packages(),
     # description='data processing module',
     zip_safe=False,
