+++ shed_streaming/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.7.5'
+__version__ = '0.8.0'
 
 try:
     from .simple_parallel import *
