### 2022-02-28 041e50d
Updated CHANGELOG for 0.8.0
