+++ news/news.rst
@@ -1,29 +1,0 @@
-**Added:**
-
-* <news item>
-
-**Changed:**
-
-* The project name is changed to `shed-streaming`.
-
-* The package name is changed to `shed_streaming`.
-
-* Update the `simple_parallel` tests using asyncio and latest dask utilities.
-
-* Pin the bluesky version to `1.6.4`.
-
-**Deprecated:**
-
-* <news item>
-
-**Removed:**
-
-* <news item>
-
-**Fixed:**
-
-* <news item>
-
-**Security:**
-
-* <news item>
