+++ CHANGELOG.rst
@@ -4,6 +4,21 @@ shed-streaming Change Log
 
 .. current developments
 
+v0.8.0
+====================
+
+**Changed:**
+
+* The project name is changed to `shed-streaming`.
+
+* The package name is changed to `shed_streaming`.
+
+* Update the `simple_parallel` tests using asyncio and latest dask utilities.
+
+* Pin the bluesky version to `1.6.4`.
+
+
+
 v0.7.5
 ====================
 
