### 2022-02-28 84bc06c
MNT: use project name shed-streaming in setup.
