+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name='shed_streaming',
+    name='shed-streaming',
     version='0.8.0',
     packages=find_packages(),
     # description='data processing module',
