### 2022-02-28 8647000
Updated CHANGELOG for 0.8.0
