+++ CHANGELOG.rst
@@ -4,10 +4,6 @@ shed-streaming Change Log
 
 .. current developments
 
-v0.8.0
-====================
-
-
 
 v0.8.0
 ====================
