+++ requirements/run.txt
@@ -4,7 +4,7 @@ numpy
 scipy
 matplotlib
 databroker
-bluesky
+bluesky==1.6
 event-model
 networkx
 rapidz
