### 2022-03-01 038af7d
Updated CHANGELOG for 0.8.0
