+++ paper.bib
@@ -66,3 +66,18 @@ tradition},
   year={2015},
   publisher={IEEE Computer Society}
 }
+@inproceedings{murray2013,
+author = {Murray, Derek and McSherry, Frank and Isaacs, Rebecca and Isard, Michael and Barham, Paul and Abadi, Martin},
+title = {Naiad: A Timely Dataflow System},
+booktitle = {Proceedings of the 24th ACM Symposium on Operating Systems Principles (SOSP)},
+year = {2013},
+month = {November},
+abstract = {Naiad is a distributed system for executing data parallel, cyclic dataflow programs. It offers the high throughput of batch processors, the low latency of stream processors, and the ability to perform iterative and incremental computations. Although existing systems offer some of these features, applications that require all three have relied on multiple platforms, at the expense of efficiency, maintainability, and simplicity. Naiad resolves the complexities of combining these features in one framework.
+
+A new computational model, timely dataflow, underlies Naiad and captures opportunities for parallelism across a wide class of algorithms. This model enriches dataflow computation with timestamps that represent logical points in the computation and provide the basis for an efficient, lightweight coordination mechanism.
+
+We show that many powerful high-level programming models can be built on Naiad’s low-level primitives, enabling such diverse tasks as streaming data analysis, iterative machine learning, and interactive graph mining. Naiad outperforms specialized systems in their target application domains, and its unique features enable the development of new high-performance applications.},
+publisher = {ACM},
+url = {https://www.microsoft.com/en-us/research/publication/naiad-a-timely-dataflow-system-2/},
+edition = {Proceedings of the 24th ACM Symposium on Operating Systems Principles (SOSP)},
+}
