### 2022-03-01 84cfb54
MNT: replace `SHED` with `shed-streaming`
