### 2022-03-08 #3119 @sbillinge
Dear @arfon and @remram44 .  Thanks for your review and your patience.  We have now responded to all the issues that were raised in @remram44 's initial review.     I will post our responses below.

The second issue @arfon raised was to help finding a second referee.  Do you want suggestions of names or do you want me to reach out directly to people?  Do you have input on the conflict of issue aspects I raised above?  The best balance would presumably be someone who is familiar with the event document-model but is not at Brookhaven National Lab.  I am now collecting good names from the BNL developers and I can paste them below.
