### 2022-03-08 #3119 @sbillinge
Here are responses to the @remram44 review points from above:

> - [x] General: shed is taken on PyPI, by https://github.com/Zac-HD/shed. It might be beneficial to use a unique name, such as shed-data, shed-streaming, etc to avoid conflicting in people's environments. Should the other shed decide to publish on conda-forge, things will get weird (I was surprised this was not discussed there
>
> [Shed conda-forge/staged-recipes#4804](https://github.com/conda-forge/staged-recipes/pull/4804)).
>

 _We changed the project name to `shed-streaming` and the module name to `shed_streaming`. The new package is released in the conda-forge channel._

> - [x] Installation instructions: Installing from the repo is not documented, only using the package from conda-forge. I think this should be added if people are to contribute or fork the project.
> I was able to install from the requirements/ files, but noted that pip.txt is empty, which is probably not intended?
>
>

_We deleted the pip.txt since it is not used in the installation. All dependencies are installed using conda._

_We add a bash script for users to install the package using the source repo and a section about how to use it in the installaion instructions._

> - [x] Automated tests: the tests don't run for me, possibly because pytest-tornado is the wrong version. I note that you use an additional Conda channel on Travis, [nsls2forge](https://anaconda.org/nsls2forge), which seems to be your own.

_The issue is probably related to the dask version.  We updated the tests so that the tests used asyncio together with the latest dask. Now, tests could be run without errors. Clone the repo and run the following commands to run the tests locally._

```
bash install.sh
conda activate <name of the environment>
python run_tests.py
```

_The nsls2forge is a channel that used to be maintained by the NSLS-II DSSI group. Now, the shed (renamed as shed-streaming) with all its dependencies are released on the conda-forge channel._

> - [x] Automated tests: you do use Travis for CI, you might want to include a link to the Travis build page (for example their badge) https://travis-ci.org/github/xpdAcq/SHED

_We included the link to the travis build. Our group runs out of the credits and thus the CI cannot be run now. We have applied for more credits_

> - [x] Functionality documentation: The documentation page https://xpdacq.github.io/SHED/shed.html shows up as empty except for headers. It looks like documentation exists and there is just something wrong with autodoc.

_We fixed the sphinx autodoc and published the new webpage._

> - [x] Documentation: The rest of the documentation is pretty sparse. There is little explanation of the purpose of the software; maybe some text can be copied from the paper..

_We copied the content in the paper into the introduction section of the documentation and it showed on the webpage._

> - [x] Tutorials: They build on NSLS-II data standards and the bluesky framework, which might not be widely known/used outside of the Brookhaven National Laboratory. It also uses rapidz, which is a fork of streamz from the same group. The paper mentions streamz but rapidz is actually what's used throughout the package, except for one example file (average.py).

_We changed the `streamz` to `rapidz` in the examples._

> - [x] Paper: The paper looks good, I sent some nit-picks at
> [JOSS paper: Minor grammar fixes xpdAcq/SHED#202](https://github.com/xpdAcq/SHED/pull/202)
>

_Grammar fixes have been merged._

> - [x] References / State of the field: Maybe more references could be added to modern streaming dataflow system/models such as Timely dataflows which are popular recently? In addition, I think more details on what SHED adds compared to dask and streamz (that it builds upon) or ray might be interesting?

_We added a reference to `Naiad: A Timely Dataflow System`._

_What functionalities that SHED adds to the rapidz (forked from streamz) and dask are:_

* Digest data stream in the event model schemas (https://nsls-ii.github.io/architecture-overview.html#event-based-architecture) and distribute the different data to corresponding consumer streams.
* Translate the data in a stream into the event model schemas and make it friendly to the database using the schemas.
* Serialize the data pipeline and rebuild the pipeline based on the serialized data.
* Replay the process that the data flow runs through the data pipeline after the data and the pipeline are loaded from the database._
