### 2022-03-10 c3b232f
MNT: update flake8; improve syntax
