### 2022-03-10 c6a0624
TST: fix the imports in confest
