### 2022-03-10 3fd2861
TST: update conftest; not to inspect it
