### 2022-04-07 #3119 @Midnighter
Randomly saw the call on Twitter. I'm not familiar with Bluesky but I'm working on a similar project myself so I'm happy to review this. My project is for internal use only, no intention to publish or sell so I don't think there's a conflict of interest. I will be very busy in the next two weeks but after that I can start.
