### 2022-04-08 #3119 @arfon
@Midnighter – many thanks for agreeing to review this submission! The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any questions/concerns please let me know. Once you're ready to start your review, please type the following into a new comment in this thread which will generate a reviewer checklist for you to work through:
