### 2022-08-07 xpdAcq/SHED#209 @Midnighter
JOSS Review: Authorship
It seems that @jrmlhermitte has made at least some contributions to SHED but is not among the authors of the paper. Just a brief check that this is agreed among all parties?
