### 2022-08-07 xpdAcq/SHED#216 @Midnighter
JOSS Review: Running the test suite fails
I tried to run the test suite but it fails with the following error:

```
coverage run run_tests.py
```

```
pluggy.manager.PluginValidationError: Plugin 'xonsh' for hook 'pytest_collect_file'
hookimpl definition: pytest_collect_file(file_path: pathlib.Path, path, parent)
Argument(s) {'file_path'} are declared in the hookimpl but can not be found in the hookspec
/home/moritz/.local/share/miniconda3/envs/shed/lib/python3.9/site-packages/coverage/control.py:794: CoverageWarning: No data was collected. (no-data-collected)
  self._warn("No data was collected.", slug="no-data-collected")
```
