### 2022-08-07 #3119 @Midnighter
**Contribution and authorship:** [I asked for quick feedback](https://github.com/xpdAcq/shed-streaming/issues/209) on a potential fourth author.

**Example usage:** While there are tutorials that help showcase the software's purpose, the examples seem rather contrived. The documentation could be improved by a real-world case, even though I won't be able to repeat that.

**Automated tests:** Tests do exist but [I was currently unable to run them](https://github.com/xpdAcq/shed-streaming/issues/216). I also was unable to find a build on Travis CI that would show me typical test output. So I'm actually in doubt if the package is continuously tested.

**Community guidelines:** Seem to be missing entirely, as also shown on the [community standards page](https://github.com/xpdAcq/shed-streaming/community).

**State of the field:** In my opinion, the authors [could provide a better overview](https://github.com/xpdAcq/shed-streaming/issues/217) here.

**References:** Although I'm not an expert on streaming data, the very minimalist representation of the state of the field makes me think that probably other important work needs to be cited.
