### 2022-10-01 #3119 @sbillinge
Dear Arfon,

So sorry about this.   I know the reviewers have put in a lot of work on
this so I am embarrassed, but actually, since CJ left for his job in
fintech, no-one has been maintaining this code base.   My new students have
not taken it over and so we are not in a position to make the changes and
updates, and it is probably fair to say that the package is no longer being
maintained and so it actually doesn't make sense to publish it anymore,
unless someone else decides to pick it up.    I was hoping to persuade my
current students to take it on and move it forward, which is why I didn't
send this email earlier, but they rather developed another parallel package
instead, so I think that, at least for the time being, SHED is dead and we
can withdraw the JOSS submission.    Again, I want to say I am really sorry
because of the effort put in by the reviewers (which would have been really
helpful if we had been able to respond).

Well, sorry it is not better news, and again, my apologies.

S
